+++ shaders/experience/cinema_film.glsl
@@ -1,6 +1,6 @@
+#version 300 es
 // HyprVision · Cinema Film
 // Curva S intensa + vinheta + granulação fina + warm split-tone
-#version 300 es
 precision highp float;
 in vec2 v_texcoord;
 layout(location = 0) out vec4 fragColor;