+++ shaders/experience/night.glsl
@@ -1,6 +1,6 @@
+#version 300 es
 // HyprVision · Night
 // Corte agressivo de azul + orange cast suave + brilho reduzido
-#version 300 es
 precision highp float;
 in vec2 v_texcoord;
 layout(location = 0) out vec4 fragColor;